+++ scripts/hail_batch/vep/README.md
@@ -3,5 +3,5 @@
 This runs a Hail query script in Dataproc using Hail Batch in order to run VEP on a hail matrix table. To run, use conda to install the analysis-runner, then execute the following command:
 
 ```sh
-analysis-runner --dataset tob-wgs --description "run vep" --output-dir "tob_wgs_vep/v0" --access-level main run_vep.py --script only_vep.py
+analysis-runner --dataset tob-wgs --description "run vep" --output-dir "tob_wgs_vep/v0" --access-level main run_vep.py --script run_vep.py
 ```
