+++ scripts/hail_batch/vep/run_vep.py
@@ -1,48 +1,32 @@
 #!/usr/bin/env python3
 
-"""
-Use VEP using a dataproc cluster.
-Taken from Matt Welland's script, run_vep_help.py
 
 """
+Run VEP on the hail mt
+"""
 
 
-import os
-import hailtop.batch as hb
-from analysis_runner import dataproc
 import click
+import hail as hl
+from analysis_runner import output_path
 
 
 @click.command()
-@click.option('--script', 'script', help='path to VEP main script')
-def main(script: str):
+@click.option('--input', required=True, help='Hail matrix table to run VEP on')
+def main(input: str):
     """
-    runs a script inside dataproc to execute VEP
-    :param script: str, the path to the VEP main script
+    Run vep using main.py wrapper
     """
 
-    service_backend = hb.ServiceBackend(
-        billing_project=os.getenv('HAIL_BILLING_PROJECT'),
-        bucket=os.getenv('HAIL_BUCKET'),
-    )
+    hl.init(default_reference='GRCh38')
 
-    # create a hail batch
-    batch = hb.Batch(name='run_vep_in_dataproc_cluster', backend=service_backend)
-
-    job = dataproc.hail_dataproc_job(
-        batch=batch,
-        script=script,
-        max_age='12h',
-        job_name='run_vep',
-        num_secondary_workers=20,
-        cluster_name='run vep',
-        vep='GRCh38',
-    )
-    job.cpu(2)
-    job.memory('standard')
-    job.storage('20G')
-
-    batch.run(wait=False)
+    mt = hl.read_matrix_table(input)
+    # filter to biallelic loci only
+    mt = mt.filter_rows(hl.len(mt.alleles) == 2)
+    mt = mt.filter_rows(mt.alleles[1] != '*')
+    vep = hl.vep(mt)
+    vep_path = output_path('vep95_GRCh38.mt')
+    vep.write(vep_path)
 
 
 if __name__ == '__main__':
